refactor(geo): extract shared photo-layer JS into photomap_js (#322)
